requirements.txt
streamlit
openai
numpy
pydub
pyaudio
wave
python-dotenv